+++ framework/db/mssql/PDO.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
@@ -8,8 +10,11 @@
 
 namespace yii\db\mssql;
 
+use PDOException;
+
 /**
  * This is an extension of the default PDO class of MSSQL and DBLIB drivers.
+ *
  * It provides workarounds for improperly implemented functionalities of the MSSQL and DBLIB drivers.
  *
  * @author Timur Ruziev <resurtm@gmail.com>
@@ -19,22 +24,29 @@ class PDO extends \PDO
 {
     /**
      * Returns value of the last inserted ID.
-     * @param string|null $sequence the sequence name. Defaults to null.
-     * @return string|false last inserted ID value.
+     *
+     * @param string|null $sequence The sequence name. Defaults to `null`.
+     *
+     * @return string|false Last inserted ID value.
      */
-    #[\ReturnTypeWillChange]
-    public function lastInsertId($sequence = null)
+    public function lastInsertId(string|null $sequence = null): string|false
     {
-        return $this->query('SELECT CAST(COALESCE(SCOPE_IDENTITY(), @@IDENTITY) AS bigint)')->fetchColumn();
+        $sql = <<<SQL
+        SELECT CAST(COALESCE(SCOPE_IDENTITY(), @@IDENTITY) AS bigint)
+        SQL;
+
+        $id = $this->query($sql)->fetchColumn();
+
+        return $id === null ? false : (string) $id;
     }
 
     /**
-     * Starts a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not
-     * natively support transactions.
-     * @return bool the result of a transaction start.
+     * Starts a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not natively support
+     * transactions.
+     *
+     * @return bool The result of a transaction start.
      */
-    #[\ReturnTypeWillChange]
-    public function beginTransaction()
+    public function beginTransaction(): bool
     {
         $this->exec('BEGIN TRANSACTION');
 
@@ -42,12 +54,12 @@ class PDO extends \PDO
     }
 
     /**
-     * Commits a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not
-     * natively support transactions.
-     * @return bool the result of a transaction commit.
+     * Commits a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not natively support
+     * transactions.
+     *
+     * @return bool The result of a transaction commit.
      */
-    #[\ReturnTypeWillChange]
-    public function commit()
+    public function commit(): bool
     {
         $this->exec('COMMIT TRANSACTION');
 
@@ -55,12 +67,12 @@ class PDO extends \PDO
     }
 
     /**
-     * Rollbacks a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not
-     * natively support transactions.
-     * @return bool the result of a transaction roll back.
+     * Rollbacks a transaction. It is necessary to override PDO's method as MSSQL PDO driver does not natively support
+     * transactions.
+     *
+     * @return bool The result of a transaction roll back.
      */
-    #[\ReturnTypeWillChange]
-    public function rollBack()
+    public function rollBack(): bool
     {
         $this->exec('ROLLBACK TRANSACTION');
 
@@ -70,21 +82,26 @@ class PDO extends \PDO
     /**
      * Retrieve a database connection attribute.
      *
-     * It is necessary to override PDO's method as some MSSQL PDO driver (e.g. dblib) does not
-     * support getting attributes.
+     * It is necessary to override PDO's method as some MSSQL PDO driver (for example, dblib) does not support getting
+     * attributes.
+     *
      * @param int $attribute One of the PDO::ATTR_* constants.
-     * @return mixed A successful call returns the value of the requested PDO attribute.
-     * An unsuccessful call returns null.
+     *
+     * @return mixed A successful call returns the value of the requested PDO attribute. An unsuccessful call returns
+     * `null`.
      */
-    #[\ReturnTypeWillChange]
-    public function getAttribute($attribute)
+    public function getAttribute(int $attribute): mixed
     {
+        $sql = <<<SQL
+        SELECT CAST(SERVERPROPERTY('productversion') AS VARCHAR)
+        SQL;
+
         try {
             return parent::getAttribute($attribute);
-        } catch (\PDOException $e) {
+        } catch (PDOException $e) {
             switch ($attribute) {
                 case self::ATTR_SERVER_VERSION:
-                    return $this->query("SELECT CAST(SERVERPROPERTY('productversion') AS VARCHAR)")->fetchColumn();
+                    return $this->query($sql)->fetchColumn();
                 default:
                     throw $e;
             }